+++ forge/src/main/java/earth/terrarium/adastra/client/forge/ClientPlatformUtilsImpl.java
@@ -1,17 +1,21 @@
 package earth.terrarium.adastra.client.forge;
 
 import earth.terrarium.adastra.client.ClientPlatformUtils;
+import earth.terrarium.adastra.client.dimension.ModDimensionSpecialEffects;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.resources.model.BakedModel;
 import net.minecraft.client.resources.model.ModelManager;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.level.Level;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class ClientPlatformUtilsImpl {
     public static final Map<Item, ArmorRenderer> ARMOR_RENDERERS = new HashMap<>();
+    public static final Map<ResourceKey<Level>, ModDimensionSpecialEffects> DIMENSION_RENDERERS = new HashMap<>();
 
     public static BakedModel getModel(ModelManager dispatcher, ResourceLocation id) {
         return dispatcher.getModel(id);
@@ -21,6 +25,11 @@ public class ClientPlatformUtilsImpl {
         for (Item item : items) {
             ARMOR_RENDERERS.put(item, new ArmorRenderer(texture, layer, factory));
         }
+    }
+
+    public static void registerPlanetRenderers(Map<ResourceKey<Level>, ModDimensionSpecialEffects> renderers) {
+        DIMENSION_RENDERERS.clear();
+        DIMENSION_RENDERERS.putAll(renderers);
     }
 
     public record ArmorRenderer(ResourceLocation texture, ModelLayerLocation layer,