+++ changelog.md
@@ -15,3 +15,4 @@ See https://modrinth.com/mod/ad-astra/version/1.15.7 for the full changelog.
 - Fixed diamond ore incorrect tag (#440)
 - Fixed waterlogged blocks getting destroyed in space (#437)
 - Fixed missing blocks in mars temple and lunarian tower.
+- Re-added planet sky rendering resource pack support.